add expense, add note, add and delete image

+++ app/(main)/(ids)/trip/planLocations.tsx
@@ -6,20 +6,27 @@ import {
   Pressable,
   Image,
   Alert,
+  TouchableOpacity,
 } from "react-native";
 import { useCallback, useEffect, useRef, useState } from "react";
 import { useTabStore } from "@/utils/store";
 import {
+  addPlanLocationImage,
   changeOrderPlanLocations,
   deletePlanLocationByPlanLocationId,
   deletePlanLocationImage,
   getPlanById,
   getPlanLocationsByPlanId,
+  patchNote,
+  putExpense,
 } from "@/services/plan/plan";
 import { useAuth } from "@/app/(auth)/AuthContext";
 import { set } from "date-fns";
 import { FlashList } from "@shopify/flash-list";
-import { GestureHandlerRootView } from "react-native-gesture-handler";
+import {
+  GestureHandlerRootView,
+  TextInput,
+} from "react-native-gesture-handler";
 import DraggableFlatList from "react-native-draggable-flatlist";
 import { Ionicons } from "@expo/vector-icons";
 import { router, Stack } from "expo-router";
@@ -27,18 +34,27 @@ import PlanLocationItem from "@/components/PlanLocation/PlanLocationItem";
 import BottomSheet, {
   BottomSheetBackdrop,
   BottomSheetModalProvider,
+  BottomSheetScrollView,
   BottomSheetView,
 } from "@gorhom/bottom-sheet";
+import MemberMultiselect from "@/components/Multiselect/MemberMultiselect";
+import { Member } from "@/constants/Member";
+import MemberDropdown from "@/components/Dropdowns/MemberDropdown";
+import * as ImagePicker from "expo-image-picker";
 
 type PlanLocationsProps = {
   planId: string;
   locationId: string;
   planLocationId: string;
   order: number;
-  images: string;
+  images: any;
   name: string;
   address: string;
   estimatedStartDate: string;
+  userSpenderIds?: string[];
+  amount: number | null;
+  payerId?: string;
+  note?: string;
 };
 
 const planLocations = () => {
@@ -51,6 +67,11 @@ const planLocations = () => {
   const [loading, setLoading] = useState(false);
   const [isRefreshing, setIsRefreshing] = useState(false);
   const [error, setError] = useState(null);
+
+  const [members, setMembers] = useState<Member[]>([]);
+  const [payer, setPayer] = useState<Member>({ userId: "", name: "" });
+  const [amount, setAmount] = useState<number>(0);
+  const [note, setNote] = useState<string>("");
 
   const tempAvatar =
     "https://icons-for-free.com/iff/png/512/mountains+photo+photos+placeholder+sun+icon-1320165661388177228.png";
@@ -71,10 +92,18 @@ const planLocations = () => {
               locationId: _item.locationId,
               planLocationId: _item.planLocationId,
               order: _item.order,
-              images: _item.images.length != 0 ? _item.images[0].url : "",
+              images:
+                _item.images.length != 0
+                  ? _item.images.map((item: any) => item.url)
+                  : [],
               name: _item.locationName,
               address: _item.locationAddress,
               estimatedStartDate: _item.estimatedStartDate.split("T")[0],
+              // userSpenderIds: _item.planLocationExpenses.map(
+              //   (item: any) => item.userId
+              // ),
+              note: _item.note,
+              amount: _item.amount,
             })
           );
         console.log("Filtered Data: ", filteredData);
@@ -113,9 +142,11 @@ const planLocations = () => {
 
   const bottomSheetRef = useRef<BottomSheet>(null);
   const handleOpen = (planLocationId: string) => {
-    setChosenPlanLocation(
-      plan.find((item) => item.planLocationId === planLocationId)
+    const _chosenPlanLocation = plan.find(
+      (item) => item.planLocationId === planLocationId
     );
+    setChosenPlanLocation(_chosenPlanLocation);
+    setNote(_chosenPlanLocation?.note!);
     bottomSheetRef.current?.expand();
   };
   const handleSheetChanges = useCallback((index: number) => {
@@ -153,24 +184,37 @@ const planLocations = () => {
     }
   };
 
-  const _deletePlanLocationImage = async (id: string, data: any) => {
-    try {
-      const _data = {
-        url: data,
-      };
-      const result = await deletePlanLocationImage(
-        _data,
-        session.userToken.accessToken,
-        id
-      );
-      if (result) {
-        console.log("Result: ", result);
-        setIsRefreshing(true);
-      }
-    } catch (error: any) {
-      console.log(error);
-    }
+  const _deletePlanLocationImage = async (data: string) => {
+    Alert.alert("Thông báo", "Bạn có chắc chắn muốn xóa ảnh này?", [
+      {
+        text: "Hủy",
+        onPress: () => console.log("Cancel Pressed"),
+        style: "cancel",
+      },
+      {
+        text: "Xóa",
+        onPress: async () => {
+          try {
+            const _data = {
+              url: data,
+            };
+            const result = await deletePlanLocationImage(
+              _data,
+              session.userToken.accessToken,
+              chosenPlanLocation!.planLocationId
+            );
+            if (result) {
+              console.log("Result: ", result);
+              setIsRefreshing(true);
+            }
+          } catch (error: any) {
+            console.log(error);
+          }
+        },
+      },
+    ]);
   };
+
   const _deletePlanLocationByPlanId = async (id: string) => {
     Alert.alert(
       "Delete Plan Location",
@@ -201,6 +245,92 @@ const planLocations = () => {
         },
       ]
     );
+  };
+
+  const handleAddExpense = async (id: string) => {
+    try {
+      const data = {
+        planLocationExpense: {
+          // userSpenderIds: [
+          //   {
+          //     userId: "3fa85f64-5717-4562-b3fc-2c963f66afa6",
+          //   },
+          // ],
+          userSpenderIds: members.map((item) => {
+            return { userId: item.userId };
+          }),
+          payerId: payer.userId,
+          amount: amount,
+        },
+      };
+      console.log("Data: ", data);
+      const result = await putExpense(data, session.userToken.accessToken, id);
+      if (result.data.isSuccess) {
+        console.log("Result: ", result);
+        setIsRefreshing(result.data.isSuccess);
+      }
+    } catch (error: any) {
+      console.log(error);
+    }
+  };
+
+  const handlePatchNote = async (id: string) => {
+    try {
+      const data = {
+        note: note,
+      };
+      console.log("Data: ", data);
+      const result = await patchNote(data, session.userToken.accessToken, id);
+      if (result) {
+        console.log("Result: ", result);
+        setIsRefreshing(true);
+      }
+    } catch (error: any) {
+      console.log(error);
+    }
+  };
+
+  const renderImage = ({ item }: { item: string }) => (
+    <TouchableOpacity
+      style={styles.imageContainer}
+      onLongPress={() => _deletePlanLocationImage(item)}
+    >
+      <Image source={{ uri: item }} style={styles.image} />
+    </TouchableOpacity>
+  );
+
+  const pickImage = async () => {
+    // No permissions request is necessary for launching the image library
+    let result = await ImagePicker.launchImageLibraryAsync({
+      mediaTypes: ImagePicker.MediaTypeOptions.All,
+      allowsEditing: true,
+      aspect: [4, 3],
+      quality: 1,
+    });
+
+    console.log(result);
+
+    if (!result.canceled) {
+      // setImage(result.assets[0].uri);
+      const parts = result.assets[0].uri.split("/"); // Chia đường dẫn theo dấu '/'
+      const fileName = parts[parts.length - 1]; // Lấy phần tử cuối cùng trong mảng (tên tệp)
+      const file: any = {
+        uri: result.assets[0].uri,
+        name: fileName,
+        type: "image/jpeg",
+      }
+      const _formData = new FormData();
+      _formData.append("image", file);
+      try {
+        const response = await addPlanLocationImage(_formData, session.userToken.accessToken, chosenPlanLocation!.planLocationId);
+        if (response) {
+          console.log("Response: ", response.data);
+          setIsRefreshing(true);
+        }
+      } catch (error: any) {
+        console.log(error);
+      }
+    }
   };
 
   if (loading) {
@@ -264,10 +394,20 @@ const planLocations = () => {
             renderItem={({ item, getIndex, drag }) => (
               <PlanLocationItem
                 index={getIndex()}
-                item={item}
+                item={{
+                  planId: item.planId,
+                  locationId: item.locationId,
+                  planLocationId: item.planLocationId,
+                  order: item.order,
+                  images: item.images[0],
+                  name: item.name,
+                  address: item.address,
+                  estimatedStartDate: item.estimatedStartDate,
+                  amount: item.amount,
+                }}
                 drag={drag}
                 onDelete={_deletePlanLocationByPlanId}
-                onDeleteImage={_deletePlanLocationImage}
+                // onDeleteImage={_deletePlanLocationImage}
                 onDetail={planDetail}
                 tempAvatar={tempAvatar}
                 _onDetail={handleOpen}
@@ -283,19 +423,82 @@ const planLocations = () => {
           <BottomSheet
             ref={bottomSheetRef}
             onChange={handleSheetChanges}
-            snapPoints={["30%"]}
+            snapPoints={["70%"]}
             index={-1}
             backdropComponent={renderBackDrop}
             enablePanDownToClose={true}
           >
-            <BottomSheetView style={{ flex: 1 }}>
+            <BottomSheetScrollView style={{ flex: 1 }}>
               <Text>Bottom Sheet Content</Text>
               <Text>{chosenPlanLocation?.planLocationId}</Text>
               <Text>{chosenPlanLocation?.order}</Text>
               <Text>{chosenPlanLocation?.name}</Text>
               <Text>{chosenPlanLocation?.address}</Text>
               <Text>{chosenPlanLocation?.estimatedStartDate}</Text>
-            </BottomSheetView>
+              <Text>Tham gia: </Text>
+              <MemberMultiselect
+                planId={sharedId!}
+                _values={members}
+                setValues={setMembers}
+                bearer={session.userToken.accessToken}
+                placeholder="Chọn thành viên"
+              />
+              <Text>Người trả: </Text>
+              {members.length != 0 ? (
+                <MemberDropdown
+                  planId={sharedId!}
+                  data={members}
+                  value={payer}
+                  setValue={setPayer}
+                  bearer={session.userToken.accessToken}
+                  placeholder="Chọn người trả"
+                />
+              ) : (
+                <Text>Chưa có người tham gia</Text>
+              )}
+              <TextInput
+                // style={styles.input}
+                keyboardType="numeric"
+                value={amount.toString()}
+                onChangeText={(text) => {
+                  const numericValue = text.replace(/[^0-9]/g, "");
+                  setAmount(parseFloat(numericValue));
+                }}
+                placeholder="Số tiền"
+              />
+              <TextInput
+                value={note}
+                onChangeText={(text) => setNote(text)}
+                placeholder="Ghi chú"
+              />
+              <Pressable
+                onPress={() =>
+                  handleAddExpense(chosenPlanLocation!.planLocationId)
+                }
+              >
+                <Text>Thêm chi phí</Text>
+              </Pressable>
+              <Pressable
+                onPress={() =>
+                  handlePatchNote(chosenPlanLocation!.planLocationId)
+                }
+              >
+                <Text>Thêm ghi chú</Text>
+              </Pressable>
+              <Pressable onPress={pickImage}>
+                <Text>Thêm hình ảnh</Text>
+              </Pressable>
+              {chosenPlanLocation?.images.length != 0 ? (
+                <FlashList
+                  data={chosenPlanLocation?.images}
+                  renderItem={renderImage}
+                  estimatedItemSize={100}
+                  horizontal
+                />
+              ) : (
+                <Text>No images</Text>
+              )}
+            </BottomSheetScrollView>
           </BottomSheet>
         </BottomSheetModalProvider>
       </GestureHandlerRootView>
